requirements.txt
gradio
numpy
openai_whisper
opencv_python
opencv_python_headless
Pillow
pydub
torch
